split consts from manifests
