+++ src/packages/media/media-types/repository/structure/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_MEDIA_TYPE_STRUCTURE_REPOSITORY_ALIAS = 'Umb.Repository.MediaType.Structure';
