+++ src/packages/media/media-types/repository/structure/index.ts
@@ -1,3 +1,3 @@
 export { UmbMediaTypeStructureRepository } from './media-type-structure.repository.js';
-export { UMB_MEDIA_TYPE_STRUCTURE_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_MEDIA_TYPE_STRUCTURE_REPOSITORY_ALIAS } from './constants.js';
 export * from './types.js';
