+++ src/packages/media/media-types/repository/structure/manifests.ts
@@ -1,6 +1,5 @@
+import { UMB_MEDIA_TYPE_STRUCTURE_REPOSITORY_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_MEDIA_TYPE_STRUCTURE_REPOSITORY_ALIAS = 'Umb.Repository.MediaType.Structure';
 
 const structureRepository: ManifestRepository = {
 	type: 'repository',
