+++ src/packages/media/media-types/tree/constants.ts
@@ -1,0 +1,3 @@
+export const UMB_MEDIA_TYPE_TREE_REPOSITORY_ALIAS = 'Umb.Repository.MediaType.Tree';
+export const UMB_MEDIA_TYPE_TREE_STORE_ALIAS = 'Umb.Store.MediaType.Tree';
+export const UMB_MEDIA_TYPE_TREE_ALIAS = 'Umb.Tree.MediaType';
