+++ src/packages/media/media-types/tree/folder/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_MEDIA_TYPE_FOLDER_REPOSITORY_ALIAS = 'Umb.Repository.MediaType.Folder';
