+++ src/packages/media/media-types/tree/folder/index.ts
@@ -1,2 +1,2 @@
 export { UmbMediaTypeFolderRepository } from './media-type-folder.repository.js';
-export { UMB_MEDIA_TYPE_FOLDER_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_MEDIA_TYPE_FOLDER_REPOSITORY_ALIAS } from './constants.js';
