+++ src/packages/media/media-types/workspace/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_MEDIA_TYPE_WORKSPACE_ALIAS = 'Umb.Workspace.MediaType';
