+++ src/packages/media/media/repository/item/index.ts
@@ -1,2 +1,2 @@
 export { UmbMediaItemRepository } from './media-item.repository.js';
-export { UMB_MEDIA_ITEM_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_MEDIA_ITEM_REPOSITORY_ALIAS } from './constants.js';
