+++ src/packages/media/media/repository/item/manifests.ts
@@ -1,7 +1,5 @@
+import { UMB_MEDIA_ITEM_REPOSITORY_ALIAS, UMB_MEDIA_STORE_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestItemStore, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_MEDIA_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.MediaItem';
-export const UMB_MEDIA_STORE_ALIAS = 'Umb.Store.MediaItem';
 
 const itemRepository: ManifestRepository = {
 	type: 'repository',
