+++ src/packages/media/media/repository/url/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_MEDIA_URL_REPOSITORY_ALIAS = 'Umb.Repository.Media.Url';
+export const UMB_MEDIA_URL_STORE_ALIAS = 'Umb.Store.MediaUrl';
