+++ src/packages/media/media/repository/url/index.ts
@@ -1,2 +1,2 @@
 export { UmbMediaUrlRepository } from './media-url.repository.js';
-export { UMB_MEDIA_URL_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_MEDIA_URL_REPOSITORY_ALIAS } from './constants.js';
