+++ src/packages/media/media/repository/url/manifests.ts
@@ -1,7 +1,5 @@
+import { UMB_MEDIA_URL_REPOSITORY_ALIAS, UMB_MEDIA_URL_STORE_ALIAS } from './constants.js';
 import type { ManifestItemStore, ManifestRepository } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_MEDIA_URL_REPOSITORY_ALIAS = 'Umb.Repository.Media.Url';
-export const UMB_MEDIA_URL_STORE_ALIAS = 'Umb.Store.MediaUrl';
 
 const urlRepository: ManifestRepository = {
 	type: 'repository',
