+++ src/packages/media/media/tree/constants.ts
@@ -1,0 +1,3 @@
+export const UMB_MEDIA_TREE_REPOSITORY_ALIAS = 'Umb.Repository.Media.Tree';
+export const UMB_MEDIA_TREE_STORE_ALIAS = 'Umb.Store.Media.Tree';
+export const UMB_MEDIA_TREE_ALIAS = 'Umb.Tree.Media';
