+++ src/packages/media/media/tree/index.ts
@@ -1,5 +1,5 @@
 export { UmbMediaTreeRepository } from './media-tree.repository.js';
-export { UMB_MEDIA_TREE_REPOSITORY_ALIAS, UMB_MEDIA_TREE_STORE_ALIAS, UMB_MEDIA_TREE_ALIAS } from './manifests.js';
+export { UMB_MEDIA_TREE_REPOSITORY_ALIAS, UMB_MEDIA_TREE_STORE_ALIAS, UMB_MEDIA_TREE_ALIAS } from './constants.js';
 export { UMB_MEDIA_TREE_STORE_CONTEXT } from './media-tree.store.js';
 export { UMB_MEDIA_TREE_PICKER_MODAL } from './media-tree-picker-modal.token.js';
 export { type UmbMediaTreeStore } from './media-tree.store.js';
